@@ -1,46 +1,37 @@
 class ProductsController < ApplicationController
-  # GET /products
-  # GET /products.json
   def index
     @products = Product.all
 
     respond_to do |format|
-      format.html # index.html.erb
+      format.html
       format.json { render json: @products }
     end
   end
 
-  # GET /products/1
-  # GET /products/1.json
   def show
     @product = Product.find(params[:id])
 
     respond_to do |format|
-      format.html # show.html.erb
+      format.html
       format.json { render json: @product }
     end
   end
 
-  # GET /products/new
-  # GET /products/new.json
   def new
     @product = Product.new
 
     respond_to do |format|
-      format.html # new.html.erb
+      format.html
       format.json { render json: @product }
     end
   end
 
-  # GET /products/1/edit
   def edit
     @product = Product.find(params[:id])
   end
 
-  # POST /products
-  # POST /products.json
   def create
-    @product = Product.new(params[:product])
+    @product = Product.new(product_params)
 
     respond_to do |format|
       if @product.save
@@ -53,13 +44,11 @@ class ProductsController < ApplicationController
     end
   end
 
-  # PUT /products/1
-  # PUT /products/1.json
   def update
     @product = Product.find(params[:id])
 
     respond_to do |format|
-      if @product.update_attributes(params[:product])
+      if @product.update(product_params)
         format.html { redirect_to @product, notice: 'Product was successfully updated.' }
         format.json { head :no_content }
       else
@@ -69,8 +58,6 @@ class ProductsController < ApplicationController
     end
   end
 
-  # DELETE /products/1
-  # DELETE /products/1.json
   def destroy
     @product = Product.find(params[:id])
     @product.destroy
@@ -80,4 +67,9 @@ class ProductsController < ApplicationController
       format.json { head :no_content }
     end
   end
+
+  private
+    def product_params
+      params.require(:product).permit(:name, :price)
+    end
 end
